Card product

+++ app/src/main/java/com/example/sarawan/framework/ui/product_card/adapter/ItemClickListener.kt
@@ -4,7 +4,7 @@ import com.example.sarawan.model.data.Product
 import com.example.sarawan.model.data.TypeCardEnum
 
 interface ItemClickListener {
-    fun openProductCard(productId : Int)
+    fun openProductCard(productId : Long)
     fun update(pos : Int = 0, mode : Boolean = false, type : TypeCardEnum)
     fun create(product: Product, pos: Int, type: TypeCardEnum)
 }
